+++ AccessoriesToStudioItems.KK/packages.config
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ExtensibleSaveFormat.Koikatu" version="16.8.1" targetFramework="net35" developmentDependency="true" />
+  <package id="IllusionLibs.BepInEx" version="5.4.20" targetFramework="net35" />
+  <package id="IllusionLibs.BepInEx.Harmony" version="2.9.0" targetFramework="net35" />
+  <package id="IllusionLibs.Koikatu.Assembly-CSharp" version="2019.4.27.4" targetFramework="net35" />
+  <package id="IllusionLibs.Koikatu.Assembly-CSharp-firstpass" version="2019.4.27.4" targetFramework="net35" />
+  <package id="IllusionLibs.Koikatu.UnityEngine" version="5.6.2.4" targetFramework="net35" />
+  <package id="IllusionLibs.Koikatu.UnityEngine.UI" version="5.6.2.4" targetFramework="net35" />
+  <package id="IllusionModdingAPI.KKAPI" version="1.36.0" targetFramework="net35" developmentDependency="true" />
+  <package id="KoikatuCompatibilityAnalyzer" version="1.1.0" targetFramework="net35" developmentDependency="true" />
+  <package id="Sideloader.Koikatu" version="19.3.2" targetFramework="net35" developmentDependency="true" />
+</packages>
