+++ src/AccessoriesToStudioItems.AI/packages.config
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="IllusionLibs.AIGirl.Assembly-CSharp" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.Assembly-CSharp-firstpass" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.MessagePack" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.Sirenix.Serialization" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.UniRx" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.UnityEngine.CoreModule" version="2018.2.21.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.UnityEngine.IMGUIModule" version="2018.2.21.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.UnityEngine.UI" version="2018.2.21.4" targetFramework="net46" />
+  <package id="IllusionLibs.AIGirl.UnityEngine.UIModule" version="2018.2.21.4" targetFramework="net46" />
+  <package id="IllusionLibs.BepInEx" version="5.4.20" targetFramework="net46" />
+  <package id="IllusionLibs.BepInEx.Harmony" version="2.9.0" targetFramework="net46" />
+  <package id="IllusionModdingAPI.AIAPI" version="1.36.0" targetFramework="net46" developmentDependency="true" />
+  <package id="Sideloader.AIGirl" version="19.3.2" targetFramework="net46" developmentDependency="true" />
+</packages>
