+++ src/AccessoriesToStudioItems.Core/AssemblyInfo.cs
