+++ src/AccessoriesToStudioItems.HS2/packages.config
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="IllusionLibs.BepInEx" version="5.4.22" targetFramework="net46" />
+  <package id="IllusionLibs.BepInEx.Harmony" version="2.9.0" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.Assembly-CSharp" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.Assembly-CSharp-firstpass" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.IL" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.Sirenix.Serialization" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.UniRx" version="2020.5.29.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.CoreModule" version="2018.4.11.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.IMGUIModule" version="2018.4.11.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.UI" version="2018.4.11.4" targetFramework="net46" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.UIModule" version="2018.4.11.4" targetFramework="net46" />
+  <package id="IllusionModdingAPI.HS2API" version="1.36.0" targetFramework="net46" developmentDependency="true" />
+  <package id="Sideloader.HoneySelect2" version="19.3.2" targetFramework="net46" developmentDependency="true" />
+</packages>
