+++ src/AccessoriesToStudioItems.KKS/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="ExtensibleSaveFormat.KoikatsuSunshine" version="16.8.1" targetFramework="net46" developmentDependency="true" />
+  <package id="ExtensibleSaveFormat.KoikatsuSunshine" version="19.3.2" targetFramework="net46" developmentDependency="true" />
   <package id="IllusionLibs.BepInEx" version="5.4.22" targetFramework="net46" />
   <package id="IllusionLibs.BepInEx.Harmony" version="2.9.0" targetFramework="net46" />
   <package id="IllusionLibs.KoikatsuSunshine.Assembly-CSharp" version="2021.9.17" targetFramework="net46" />
